+++ include/menu.php
@@ -215,6 +215,7 @@ if($idleto != "disable"){
 <?php 
 } ?>  
   <a href="./admin.php?id=sessions" class="menu <?= $ssesslist; ?>"><i class="fa fa-gear"></i> <?= $_admin_settings ?></a>
+  <a href="./admin.php?id=pending-transactions" class="menu <?= ($id == 'pending-transactions') ? 'active' : ''; ?>"><i class="fa fa-exchange"></i> <?= ($langid == 'id') ? 'Antrean Webhook' : 'Webhook Queue' ?></a>
   <a href="./admin.php?id=settings&router=new-<?= rand(1111,9999) ?>" class="menu <?= $snsettings ?>"><i class="fa fa-plus"></i> <?= $_add_router ?></a>
   <a href="./admin.php?id=about" class="menu <?= $sabout; ?>"><i class="fa fa-info-circle"></i> <?= $_about ?></a>